flask-a/requirements.txt
Flask
Flask-Cors
gunicorn
semgrep
python-dotenv